rm root slashes from directories
it looks like the PWA audit isn't working on the github
site. It seems that these URLs aren't resolving as they
start w/a slash, but the site is at domain/officers-log
rather than the root of the site. Trying deleting them
to see if that fixes it instead of needing to specify
site name or try other options